+++ src/app.html
@@ -3,9 +3,13 @@
     <head>
         <meta charset="UTF-8">
         <title>Museeks</title>
+
+        <!-- Bower Components -->
         <link rel="stylesheet" href="components/bootstrap-css-only/css/bootstrap.min.css">
         <link rel="stylesheet" href="components/font-awesome/css/font-awesome.min.css">
         <link rel="stylesheet" href="components/bootstrap-sharper/bootstrap-sharper.css">
+
+        <!-- Other Styles-->
         <link rel="stylesheet" href="css/main.css">
     </head>
 
@@ -19,8 +23,11 @@
         <script src="components/react-bootstrap/react-bootstrap.min.js"></script>
 
         <!-- App -->
+        <script src="js/functions.js"></script>
         <script src="js/alerts.js"></script>
         <script src="js/app.js"></script>
+
+
 
         <!-- React Components -->
         <script src="js/components/bootstrap.js"></script>
