+++ src/js/routes.js
@@ -4,9 +4,5 @@ routie({
     },
     '/settings': function() {
         Instance.setState({view: views.settings});
-    },
-    '*': function() {
-        alerts.add('warning', 'You tried to access an undefined view !');
-        routie('/');
     }
 });
